requirements.txt
commitizen>=3.12.0
typer>=0.9.0
rich>=13.7.0
google-generativeai>=0.3.2
absl-py>=2.0.0
questionary>=2.0.1 